Cleaned up Preparation section a bit
Added some of the required tools (as far as I've needed them up to now). 
Added some details to the parts lists as far as I could from my own kit.
Improved the footer secion to include links to the mellow kit STL zipped archive and the Printed Parts guide to point people to the print settings.
Fixed some typos.